+++ tests/test_series.py
@@ -2608,6 +2608,67 @@ def test_sliced_vector_series_slice_instantiation():
     assert isinstance(s.slice_surf_series, SurfaceInteractiveSeries)
 
 
+def test_is_polar_2d_parametric():
+    # verify that Parametric2DLineSeries and Parametric2DInteractiveLineSeries
+    # are able to apply polar discretization, which is used when polar_plot
+    # is executed with polar_axis=True
+    t, u = symbols("t u")
+
+    # NOTE: a sufficiently big n must be provided, or else tests
+    # are going to fail
+    # No colormap
+    f = sin(4 * t)
+    s1 = Parametric2DLineSeries(f * cos(t), f * sin(t), (t, 0, 2*pi),
+        adaptive=False, n=10, is_polar=False, use_cm=False)
+    x1, y1, p1 = s1.get_data()
+    s2 = Parametric2DLineSeries(f * cos(t), f * sin(t), (t, 0, 2*pi),
+        adaptive=False, n=10, is_polar=True, use_cm=False)
+    th, r, p2 = s2.get_data()
+    assert (not np.allclose(x1, th)) and (not np.allclose(y1, r))
+    assert np.allclose(p1, p2)
+
+    # With colormap
+    s3 = Parametric2DLineSeries(f * cos(t), f * sin(t), (t, 0, 2*pi),
+        adaptive=False, n=10, is_polar=False, color_func=lambda t: 2*t)
+    x3, y3, p3 = s3.get_data()
+    s4 = Parametric2DLineSeries(f * cos(t), f * sin(t), (t, 0, 2*pi),
+        adaptive=False, n=10, is_polar=True, color_func=lambda t: 2*t)
+    th4, r4, p4 = s4.get_data()
+    assert np.allclose(p3, p4) and (not np.allclose(p1, p3))
+    assert np.allclose(x3, x1) and np.allclose(y3, y1)
+    assert np.allclose(th4, th) and np.allclose(r4, r)
+
+    # No colormap
+    s1i = Parametric2DLineInteractiveSeries(
+        [f * cos(t), f * sin(t)], [(t, 0, 2*pi)], params={u: 1},
+        adaptive=False, n1=10, is_polar=False, use_cm=False)
+    x1i, y1i, p1i = s1i.get_data()
+    s2i = Parametric2DLineInteractiveSeries(
+        [f * cos(t), f * sin(t)], [(t, 0, 2*pi)], params={u: 1},
+        adaptive=False, n1=10, is_polar=True, use_cm=False)
+    thi, ri, p2i = s2i.get_data()
+    assert (not np.allclose(x1i, thi)) and (not np.allclose(y1i, ri))
+    assert np.allclose(x1, x1i) and np.allclose(y1, y1i)
+    assert np.allclose(th, thi) and np.allclose(r, ri)
+    assert np.allclose(p1i, p2i)
+
+    # With colormap
+    s3i = Parametric2DLineInteractiveSeries(
+        [f * cos(t), f * sin(t)], [(t, 0, 2*pi)], params={u: 1},
+        adaptive=False, n1=10, is_polar=False, color_func=lambda t: 2*t)
+    x3i, y3i, p3i = s3i.get_data()
+    s4i = Parametric2DLineInteractiveSeries(
+        [f * cos(t), f * sin(t)], [(t, 0, 2*pi)], params={u: 1},
+        adaptive=False, n1=10, is_polar=True, color_func=lambda t: 2*t)
+    th4i, r4i, p4i = s4i.get_data()
+    assert np.allclose(p3i, p4i)
+    assert np.allclose(x3i, x1i) and np.allclose(y3i, y1i)
+    assert np.allclose(th4i, thi) and np.allclose(r4i, ri)
+    assert np.allclose(p3i, p3) and np.allclose(p4i, p4)
+    assert np.allclose(x3i, x3) and np.allclose(y3i, y3)
+    assert np.allclose(th4i, th4) and np.allclose(r4i, r4)
+
+
 def test_is_polar_3d():
     # verify that SurfaceOver2DRangeSeries and SurfaceInteractiveSeries
     # are able to apply polar discretization